requirements.txt
aniso8601==9.0.1
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
backports.zoneinfo==0.2.1
cffi==1.16.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.95.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.14
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.11
starlette==0.26.1
starlette-graphene3==0.6.0
typing_extensions==4.10.0
uvicorn==0.21.1
